Remove PageListService
