+++ packages/sitecore-jss/src/graphql/index.ts
@@ -1,5 +1,4 @@
 export { getAppRootId, AppRootQueryResult } from './app-root-query';
-export { PageListService, PageListQueryResult } from './page-list-service';
 export {
   SearchQueryResult,
   SearchQueryVariables,
